requirements.txt
skimage
tensorflow==2.5.0
pillow
sklearn
matplotlib
seaborn
pandas